+++ core/src/main/java/tech/ydb/core/auth/AuthIdentity.java
@@ -1,0 +1,14 @@
+package tech.ydb.core.auth;
+
+/**
+ * @author Sergey Polovko
+ */
+public interface AuthIdentity extends AutoCloseable {
+
+    String getToken();
+
+    @Override
+    default void close() {
+    }
+}
+